Moved peers rpc server to rpcservers folder
