+++ log.go
@@ -21,7 +21,6 @@ import (
 	"github.com/lightningnetwork/lnd/healthcheck"
 	"github.com/lightningnetwork/lnd/htlcswitch"
 	"github.com/lightningnetwork/lnd/invoices"
-	"github.com/lightningnetwork/lnd/lnrpc/peersrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/routerrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/signrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/verrpc"
@@ -42,6 +41,7 @@ import (
 	devrpc "github.com/lightningnetwork/lnd/rpcservers/dev"
 	invoicesrpc "github.com/lightningnetwork/lnd/rpcservers/invoices"
 	neutrinorpc "github.com/lightningnetwork/lnd/rpcservers/neutrino"
+	peersrpc "github.com/lightningnetwork/lnd/rpcservers/peers"
 	"github.com/lightningnetwork/lnd/signal"
 	"github.com/lightningnetwork/lnd/sweep"
 	"github.com/lightningnetwork/lnd/tor"
