+++ rpcservers/peers/config_default.go
@@ -1,7 +1,7 @@
 //go:build !peersrpc
 // +build !peersrpc
 
-package peersrpc
+package peers
 
 // Config is empty for non-peersrpc builds.
 type Config struct{}
