+++ subrpcserver_config.go
@@ -13,7 +13,6 @@ import (
 	"github.com/lightningnetwork/lnd/htlcswitch"
 	"github.com/lightningnetwork/lnd/invoices"
 	"github.com/lightningnetwork/lnd/lncfg"
-	"github.com/lightningnetwork/lnd/lnrpc/peersrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/routerrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/signrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/walletrpc"
@@ -28,6 +27,7 @@ import (
 	devrpc "github.com/lightningnetwork/lnd/rpcservers/dev"
 	invoicesrpc "github.com/lightningnetwork/lnd/rpcservers/invoices"
 	neutrinorpc "github.com/lightningnetwork/lnd/rpcservers/neutrino"
+	peersrpc "github.com/lightningnetwork/lnd/rpcservers/peers"
 	"github.com/lightningnetwork/lnd/sweep"
 	"github.com/lightningnetwork/lnd/watchtower"
 	"github.com/lightningnetwork/lnd/watchtower/wtclient"
